App layout: Add appstream download button
Also adds an optional, off-by-default .flatpakref download button if a
`remote_url` exists in the site config. The idea being that we could add
direct downloads of .flatpakrefs if they were added in the future.

Unfortunately after writing this I learned that we've never created
.flatpakref files for eos-apps, and that's outside the current scope of
work. But it's here if that ever changes (and may be useful to
elementary's AppCenter site upstream).

Mixed in a bit are some code legibility and path fixes I noticed while
developing locally.

https://phabricator.endlessm.com/T34894